New: Callback of discovery interface in separate callback group
The DiscoveryInterface allows to pass an "on_activate" function that gets called when a new entity is detected.
This function gets called from a callback.
This means that if "on_activate" does other calls such as a service call, this call will cause a deadlock as we can not process the reply [1].

This design allows for arbitrary procedures in "on_active".

[1] https://docs.ros.org/en/foxy/How-To-Guides/Using-callback-groups.html#the-problem